+++ controls-constructor/src/commonMain/kotlin/space/kscience/controls/constructor/models/Reducer.kt
@@ -5,6 +5,7 @@ import space.kscience.controls.constructor.units.Degrees
 import space.kscience.controls.constructor.units.NumericAmount
 import space.kscience.controls.constructor.units.times
 import space.kscience.dataforge.context.Context
+import space.kscience.dataforge.names.asName
 
 /**
  * A reducer device used for simulations only (no public properties)
@@ -16,8 +17,8 @@ public class Reducer(
     public val output: MutableValueState<NumericAmount<Degrees>>,
 ) : ModelConstructor(context) {
     init {
-        registerState(input)
-        registerState(output)
+        registerState(input, "input".asName())
+        registerState(output, "output".asName())
         bindTransformedState(input, output) {
             it * ratio
         }