+++ controls-constructor/src/commonMain/kotlin/space/kscience/controls/constructor/models/continuous/ContinuousBuffer.kt
@@ -3,6 +3,9 @@ package space.kscience.controls.constructor.models.continuous
 import space.kscience.controls.constructor.*
 import space.kscience.controls.constructor.units.*
 import space.kscience.dataforge.context.Context
+import space.kscience.dataforge.names.Name
+import space.kscience.dataforge.names.asName
+import space.kscience.dataforge.names.parseAsName
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.seconds
 
@@ -18,7 +21,7 @@ import kotlin.time.Duration.Companion.seconds
  * @param supplyRequest A device state representing the amount requested to be added to the buffer.
  * @param consumerRequest A device state representing the amount requested to be consumed from the buffer.
  * @param initialLevel The initial quantity stored in the buffer.
- * @param timeStep The simulation time step for updating the buffer's state.
+ * @param externalTimer An optional timer to be used for dynamic level change calculations.
  *
  * The model tracks the current level of stored content, calculates remaining buffer space, and manages production and consumption dynamics.
  * It ensures the buffer's level stays within the bounds defined by its capacity.
@@ -31,7 +34,7 @@ public class ContinuousBuffer<U : UnitsOfMatter, T : Amount<U>>(
     override val consumerRequest: LateBindValueState<AmountPerSecond<U>> = LateBindValueState(AmountPerSecond.zero()),
     initialLevel: T = consumerAlgebra.zero,
     externalTimer: TimerState? = null,
-) : ModelConstructor(context, bufferCapacity, supplyRequest, consumerRequest),
+) : ModelConstructor(context),
     ContinuousProducer<U, T>,
     ContinuousConsumer<U, T> {
 
@@ -42,12 +45,16 @@ public class ContinuousBuffer<U : UnitsOfMatter, T : Amount<U>>(
     public val content: ValueState<T> get() = _content
 
     init {
-        registerState(content)
+        registerState(content, "content".asName())
+        registerState(bufferCapacity, "bufferCapacity".asName())
+        registerState(supplyRequest, "supply.request".parseAsName(true))
+        registerState(consumerRequest, "consumer.request".parseAsName(true))
     }
 
     override val productionCapacity: ValueState<PerSecond<U, T>> = combineState(
-        supplyRequest,
-        content
+        first = supplyRequest,
+        second = content,
+        name = "production.capacity".parseAsName(true)
     ) { supplyRequest: PerSecond<U, T>, content: T ->
         with(consumerAlgebra) {
             supplyRequest + content.perSecond
@@ -55,9 +62,10 @@ public class ContinuousBuffer<U : UnitsOfMatter, T : Amount<U>>(
     }
 
     override val production: ValueState<PerSecond<U, T>> = combineState(
-        supplyRequest,
-        content,
-        consumerRequest
+        first = supplyRequest,
+        second = content,
+        third = consumerRequest,
+        name = "production".asName()
     ) { supplyRequest, content, consumeRequest ->
         with(consumerAlgebra) {
             val productionCapacity = supplyRequest + content.perSecond
@@ -66,18 +74,20 @@ public class ContinuousBuffer<U : UnitsOfMatter, T : Amount<U>>(
     }
 
     override val consumationCapacity: ValueState<AmountPerSecond<U>> = combineState(
-        bufferCapacity,
-        content,
-        consumerRequest,
+        first = bufferCapacity,
+        second = content,
+        third = consumerRequest,
+        name = "consumation.capacity".parseAsName(true)
     ) { bufferSize: NumericAmount<U>, content: T, consumationRequest: AmountPerSecond<U> ->
         AmountPerSecond(consumationRequest.value + bufferSize.value - content.value)
     }
 
     override val consumation: ValueState<PerSecond<U, T>> = combineState(
-        supplyRequest,
-        bufferCapacity,
-        content,
-        consumerRequest
+        first = supplyRequest,
+        second = bufferCapacity,
+        third = content,
+        forth = consumerRequest,
+        name = "consumation".asName()
     ) { supplyRequest: PerSecond<U, T>, bufferCapacity: NumericAmount<U>, content: T, consumationRequest ->
         with(consumerAlgebra) {
             val remainingSpace = bufferCapacity - NumericAmount<U>(content.value)
@@ -114,8 +124,6 @@ public class ContinuousBuffer<U : UnitsOfMatter, T : Amount<U>>(
  *
  * @param context The simulation context in which the buffer operates.
  * @param capacity The maximum capacity of the buffer, represented as a [PerSecond] value.
- * @param supplyRequest A device state representing the amount requested to be supplied
- *        to the buffer. Defaults to a [LateBindValueState] with a value of zero.
  * @return A [ContinuousBuffer] instance configured with the specified parameters.
  */
 public fun <U : UnitsOfMatter> ContinuousBuffer(
@@ -132,6 +140,7 @@ public fun <U : UnitsOfMatter, T : Amount<U>> ContinuousFlowModel.buffer(
     bufferCapacity: NumericAmount<U>,
     initialLevel: T = algebra.zero,
     externalTimer: TimerState? = null,
+    modelName: Name? = null
 ): ContinuousBuffer<U, T> = model(
     ContinuousBuffer(
         context = context,
@@ -139,5 +148,6 @@ public fun <U : UnitsOfMatter, T : Amount<U>> ContinuousFlowModel.buffer(
         bufferCapacity = ValueState(bufferCapacity),
         initialLevel = initialLevel,
         externalTimer = externalTimer,
-    )
+    ),
+    modelName
 )