+++ controls-constructor/src/commonMain/kotlin/space/kscience/controls/constructor/models/continuous/ContinuousFlowModel.kt
@@ -1,15 +1,13 @@
 package space.kscience.controls.constructor.models.continuous
 
 import space.kscience.controls.constructor.ModelConstructor
-import space.kscience.controls.constructor.ValueState
 import space.kscience.controls.constructor.units.Amount
 import space.kscience.controls.constructor.units.UnitsOfMatter
 import space.kscience.dataforge.context.Context
 
 public abstract class ContinuousFlowModel(
     context: Context,
-    vararg dependencies: ValueState<*>
-) : ModelConstructor(context, *dependencies) {
+) : ModelConstructor(context) {
 
     public companion object {
         public fun <U : UnitsOfMatter, T : Amount<U>> connect(