+++ controls-constructor/src/commonMain/kotlin/space/kscience/controls/constructor/models/continuous/ContinuousTransformer.kt
@@ -3,6 +3,9 @@ package space.kscience.controls.constructor.models.continuous
 import space.kscience.controls.constructor.*
 import space.kscience.controls.constructor.units.*
 import space.kscience.dataforge.context.Context
+import space.kscience.dataforge.names.Name
+import space.kscience.dataforge.names.asName
+import space.kscience.dataforge.names.parseAsName
 
 public interface ContinuousTransformationRule<U1 : UnitsOfMatter, T : Amount<U1>, U2 : UnitsOfMatter, R : Amount<U2>> {
     public fun computeProduction(amount: PerSecond<U1, T>): PerSecond<U2, R>
@@ -45,16 +48,25 @@ public class ContinuousTransformer<U1 : UnitsOfMatter, T : Amount<U1>, U2 : Unit
 
     override val supplyRequest: LateBindValueState<PerSecond<U1, T>> =
         LateBindValueState(consumerAlgebra.zero.perSecond)
+
     override val consumerRequest: LateBindValueState<AmountPerSecond<U2>> = LateBindValueState(PerSecond.zero())
 
+    init {
+        registerState(supplyRequest, "supply.request".parseAsName(true))
+        registerState(consumerRequest, "consumer.request".parseAsName(true))
+    }
+
     override val consumation: ValueState<PerSecond<U1, T>> =
-        combineState(supplyRequest, consumerRequest) { supply, consume ->
+        combineState(supplyRequest, consumerRequest, "consumation".asName()) { supply, consume ->
             with(consumerAlgebra) {
                 supply.coerceValueIn(consumerAlgebra.zero..rule.computeConsumption(consume))
             }
         }
 
-    override val consumationCapacity: ValueState<AmountPerSecond<U1>> = mapState(consumerRequest) {
+    override val consumationCapacity: ValueState<AmountPerSecond<U1>> = mapState(
+        origin = consumerRequest,
+        name = "consumation.capacity".parseAsName(true)
+    ) {
         rule.computeConsumption(it)
     }
 
@@ -69,17 +81,20 @@ public fun <U1 : UnitsOfMatter, T : Amount<U1>, U2 : UnitsOfMatter, R : Amount<U
     consumerAlgebra: AmountAlgebra<U1, T>,
     producerAlgebra: AmountAlgebra<U2, R>,
     rule: ContinuousTransformationRule<U1, T, U2, R>,
-): ContinuousTransformer<U1, T, U2, R> = model(ContinuousTransformer(context, consumerAlgebra, producerAlgebra, rule))
+    modelName: Name? = null
+): ContinuousTransformer<U1, T, U2, R> = model(ContinuousTransformer(context, consumerAlgebra, producerAlgebra, rule), modelName)
 
 public fun <U1 : UnitsOfMatter, T : Amount<U1>, U2 : UnitsOfMatter, R : Amount<U2>> ContinuousFlowModel.linearTransformer(
     consumerAlgebra: AmountAlgebra<U1, T>,
     producerAlgebra: AmountAlgebra<U2, R>,
-    production: PerSecond<U2, R>
+    production: PerSecond<U2, R>,
+    modelName: Name? = null
 ): ContinuousTransformer<U1, T, U2, R> = model(
     ContinuousTransformer(
         context = context,
         producerAlgebra = producerAlgebra,
         consumerAlgebra = consumerAlgebra,
         rule = ContinuousTransformationRule.linear(producerAlgebra, production)
-    )
+    ),
+    modelName
 )